Extract DynamoDBStateStoreTestBase
