+++ java/statestore/src/test/java/sleeper/statestore/dynamodb/DynamoDBStateStoreTestBase.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2022-2023 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sleeper.statestore.dynamodb;
+
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+import sleeper.configuration.properties.instance.InstanceProperties;
+import sleeper.dynamodb.tools.DynamoDBContainer;
+
+import static sleeper.configuration.properties.InstancePropertiesTestHelper.createTestInstanceProperties;
+import static sleeper.dynamodb.tools.GenericContainerAwsV1ClientHelper.buildAwsV1Client;
+
+@Testcontainers
+public class DynamoDBStateStoreTestBase {
+    protected static AmazonDynamoDB dynamoDBClient;
+
+    @Container
+    public static DynamoDBContainer dynamoDb = new DynamoDBContainer();
+
+    protected final InstanceProperties instanceProperties = createTestInstanceProperties();
+
+    @BeforeAll
+    public static void initDynamoClient() {
+        dynamoDBClient = buildAwsV1Client(dynamoDb, dynamoDb.getDynamoPort(), AmazonDynamoDBClientBuilder.standard());
+    }
+
+    @AfterAll
+    public static void shutdownDynamoClient() {
+        dynamoDBClient.shutdown();
+    }
+
+    @BeforeEach
+    void setUp() {
+        new DynamoDBStateStoreCreator(instanceProperties, dynamoDBClient).create();
+    }
+}
